Update fct logging

+++ common/identity/entryBlockSync.go
@@ -143,7 +143,7 @@ func (e *EntryBlockSync) UnmarshalBinaryData(p []byte) (newData []byte, err erro
 	// TODO: remove printing unmarshal count numbers once we have good data on
 	// what they should be.
 	//log.Print("EntryBlockSync unmarshaled EntryBlockMarker count: ", l)
-	if l > 1000 {
+	if uint(l) > 1000 {
 		// TODO: replace this message with a proper error
 		return nil, fmt.Errorf("Error: EntryBlockSync.UnmarshalBinary: EntryBlockMarker count too high (uint underflow?)")
 	}
